% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chords.R
\name{as.character.pichor_chord}
\alias{as.character.pichor_chord}
\title{Get textual description of chord}
\usage{
\method{as.character}{pichor_chord}(x, brief = FALSE, ...)
}
\arguments{
\item{x}{a \code{pichor_chord} from e.g. \code{\link[=construct_chord_raw]{construct_chord_raw()}},
\code{\link[=construct_chord_major]{construct_chord_major()}} or \code{\link[=construct_chord_minor]{construct_chord_minor()}}}

\item{brief}{brief description}

\item{\dots}{currently not used}
}
\description{
Get textual description of chord
}
