% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chords.R
\name{construct_chord_dominant_7}
\alias{construct_chord_dominant_7}
\title{Get dominant 7th of chord}
\usage{
construct_chord_dominant_7(root_tone)
}
\arguments{
\item{root_tone}{root tone of chord}
}
\description{
Get dominant 7th of chord
}
