% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chords.R
\name{construct_chord_major}
\alias{construct_chord_major}
\title{Get major of chord}
\usage{
construct_chord_major(root_tone)
}
\arguments{
\item{root_tone}{root tone of chord}
}
\description{
Get major of chord
}
