% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chords.R
\name{construct_chord_minor}
\alias{construct_chord_minor}
\title{Get minor of chord}
\usage{
construct_chord_minor(root_tone)
}
\arguments{
\item{root_tone}{root tone of chord}
}
\description{
Get minor of chord
}
