% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chords.R
\name{construct_chord_raw}
\alias{construct_chord_raw}
\title{Construct a chord by root tone and key distances}
\usage{
construct_chord_raw(root_tone, distances_rel)
}
\arguments{
\item{root_tone}{root tone of chord}

\item{distances_rel}{relative distances: major is e.g. \code{c(4, 3)}, minor \code{c(3, 4)} etc.}
}
\description{
Construct a chord by root tone and key distances
}
