% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_keys_chords}
\alias{generate_keys_chords}
\title{Generate a custom-sized piano}
\usage{
generate_keys_chords(number_white_keys = 14)
}
\description{
Generate a custom-sized piano, useful e.g. for synthesizers.
}
\examples{
d <- generate_keys_chords(10)
ggpiano(d, labels = TRUE)
d <- d \%>\% 
  dplyr::rowwise() \%>\% 
  dplyr::mutate(label = tones[[1]])
ggpiano(d, labels = TRUE)
 
}
