% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chords.R
\name{get_keys}
\alias{get_keys}
\title{Get keys for a chord}
\usage{
get_keys(chord)
}
\arguments{
\item{chord}{chord created by e.g. \verb{construct_chord_*()}, i.e. \code{\link[=construct_chord_raw]{construct_chord_raw()}},
\code{\link[=construct_chord_major]{construct_chord_major()}}, \code{\link[=construct_chord_minor]{construct_chord_minor()}}}
}
\description{
Get keys for a chord
}
