% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chords.R
\name{get_keys_highest_tone}
\alias{get_keys_highest_tone}
\title{Get keys for an inversion with a specific highest tone}
\usage{
get_keys_highest_tone(chord, highest_tone)
}
\arguments{
\item{chord}{chord created by e.g. \verb{construct_chord_*()}, i.e. \code{\link[=construct_chord_raw]{construct_chord_raw()}},
\code{\link[=construct_chord_major]{construct_chord_major()}}, \code{\link[=construct_chord_minor]{construct_chord_minor()}}}

\item{highest_tone}{Higest tone in the inversion wanted to get keys for}
}
\description{
Get keys for an inversion with a specific highest tone
}
