% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chords.R
\name{get_next_key}
\alias{get_next_key}
\title{Get next key}
\usage{
get_next_key(origin, distance)
}
\arguments{
\item{origin}{origin key}

\item{distance}{number of keys to move}
}
\description{
Get next key
}
