% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chords.R
\name{get_tones}
\alias{get_tones}
\title{Get tones for a key}
\usage{
get_tones(key, sep = "/", as_string = TRUE)
}
\arguments{
\item{key}{key to get tone(s) for}

\item{sep}{when returning string, chactecter to concatenate with}

\item{as_string}{whether to return tones as string concatenated with \code{sep} or as a vector}
}
\description{
Get tones for a key
}
