% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggchord}
\alias{ggchord}
\title{Add chord}
\usage{
ggchord(chord = NULL, color = "lightblue", data = keys_coords)
}
\arguments{
\item{chord}{chord to plot}

\item{color}{color for keys in chord}

\item{data}{data to plot, default \code{keys_coords} which is a piano}
}
\description{
Add chord
}
