% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{highlight_chord}
\alias{highlight_chord}
\title{Highlight coord}
\usage{
highlight_chord(data, chord, inversion = 0L, highest_tone = NULL,
  color = "lightblue")
}
\arguments{
\item{data}{data with key coordinates, e.g. from \link{keys_coords}}

\item{chord}{chord to highlight, e.g. from \code{\link[=construct_chord_major]{construct_chord_major()}} or \code{\link[=construct_chord_raw]{construct_chord_raw()}}}

\item{inversion}{inversion number}

\item{highest_tone}{if inversion by highest tone is wanted}

\item{color}{highlight color}
}
\description{
Highlight coord
}
\details{
If \code{highest_tone} is provided, then \code{inversion} is ignored
}
