% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{highlight_key_sequence}
\alias{highlight_key_sequence}
\title{Highlight key sequence}
\usage{
highlight_key_sequence(data, key_sequence, sequence_names = NULL,
  new_color = "lightblue", keep_color = "lightgrey",
  remove_color = NULL)
}
\arguments{
\item{data}{data with key coordinates, e.g. from \link{keys_chords}}

\item{key_sequence}{list of vector with key numbers, note that they can be greater than 12}

\item{sequence_names}{names of each element in \code{key_sequence}}

\item{new_color}{highlight color for new keys (\code{NULL} for disable)}

\item{keep_color}{highlight color for keys kept at same position (\code{NULL} for disable)}

\item{remove_color}{highlight color for removed keys (\code{NULL} for disable)}
}
\value{
A new dataset with two new columns: \code{seq_no}, that can be used in e.g. plotting,
and \code{seq_name} that is either \code{seq_no} of \code{sequence_names} if provided.
}
\description{
Highlight key sequence
}
