% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{highlight_keys}
\alias{highlight_keys}
\title{Highlight keys}
\usage{
highlight_keys(data, keys, color = "lightblue")
}
\arguments{
\item{data}{data with key coordinates, e.g. from \code{\link[=get_keys_coords]{get_keys_coords()}}}

\item{keys}{key numbers, note that can be greater than 12}

\item{color}{highlight color}
}
\description{
Highlight keys
}
