% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{keys_coords}
\alias{keys_coords}
\title{Coordinates to draw a piano}
\format{A tibble containing information
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}}
\usage{
keys_coords
}
\description{
Coordinates to draw a piano
}
\keyword{datasets}
