% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{optim_min_dist_exhaustive}
\alias{optim_min_dist_exhaustive}
\title{Optimise sequence}
\usage{
optim_min_dist_exhaustive(key_sequence, error_on_big = TRUE)
}
\arguments{
\item{key_sequence}{list of vector with key numbers, note that they can be greater than 12}

\item{error_on_big}{stop if more than 100,000 combinations of inversions must be checked}
}
\description{
Exhaustive minimisation of the distances between keys in consecutive chords
}
