% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggpiano}
\alias{ggpiano}
\title{Plot piano}
\usage{
ggpiano(data = keys_chords, labels = FALSE, ...)
}
\arguments{
\item{data}{data to plot, default \code{keys_chords} which is a piano}

\item{labels}{whether to show key labels}

\item{\dots}{Currently not used}
}
\description{
Plot piano
}
