% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{highlight_key_sequence}
\alias{highlight_key_sequence}
\title{Highlight key sequence}
\usage{
highlight_key_sequence(data, key_sequence, new_color = "lightblue",
  keep_color = "lightgrey")
}
\arguments{
\item{data}{data with key coordinates, e.g. from \link{keys_coords}}

\item{key_sequence}{list of vector with key numbers, note that they can be greater than 12}

\item{new_color}{highlight color for new keys}

\item{keep_color}{highlight color for keys kept at same position}
}
\value{
A new dataset with a new column, \code{seq_no}, that can be used in e.g. plotting
}
\description{
Highlight key sequence
}
